Move general models to a folder
