+++ api/serializers/action.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from general.models import Action
+from general.models.action import Action
 
 
 class ActionsSerializer(serializers.ModelSerializer):
