+++ course/models/models.py
@@ -15,7 +15,7 @@ from course.fields import JSONField
 from course.utils.junit_xml import parse_junit_xml
 from course.utils.utils import get_token_value, ensure_uqj, calculate_average_success
 from course.utils.variables import render_text, generate_variables
-from general.models import Action
+from general.models.action import Action
 
 DIFFICULTY_CHOICES = [
     ("EASY", "EASY"),
